rename metadata.ts to use-intercom-metadata.ts
