+++ src/chat/use-intercom-metadata.ts
